Re-wrote and mostly fixed MIDI loop scheduler
 * I really hate these schedulers.  They're such a huge task to implement; have to convert between like 4 different spaces and units and handle 3 offsets at once in order to implement.  Ugh.
 * I basically tore out the whole old one and re-built it from scratch with extra-verbose variable names and a lot of debugging time.
 * Replaced all instances of `Box::new()` with `box` syntax because apparently `rust-analyzer` supports that now

+++ engine/engine/src/js.rs
@@ -113,16 +113,16 @@ extern "C" {
         vc_id: &str,
         events: &[u8],
         note_ids: &[usize],
-        timings: &[f32],
+        timings: &[f64],
     );
     pub fn midi_editor_cancel_all_events(vc_id: &str);
     pub fn register_midi_editor_loop_interval(
-        cb: &Closure<dyn FnMut(f32)>,
+        cb: &Closure<dyn FnMut(f64)>,
         inteval_ms: usize,
     ) -> SchedulerLoopHandle;
     pub fn cancel_midi_editor_loop_interval(handle: SchedulerLoopHandle);
     pub fn midi_editor_register_animation_frame(
-        cb: &Closure<dyn FnMut(f32)>,
+        cb: &Closure<dyn FnMut(f64)>,
     ) -> SchedulerLoopHandle;
     pub fn midi_editor_cancel_animation_frame(handle: SchedulerLoopHandle);
 }
